Major update!
Created a new class, BoardData.java to hold the board's data and the
class implements Comparable so it can be used in the priority queue as
part of the A* algorithm for solving the puzzle.  Added a ButtonGUI.java
class to add an additional window containing buttons for the user to
interact with.  Added logic to the NPuzzleGUI.java class that uses a
priority queue for the A* algorithm and uses the Manhattan distance for
the priority key value.  I'm still working on finishing out the button
GUI.